Add 0.15.13 as supported version & changes to SW.README
Replace � around fair use with " and I think that's what they were supposed to be. There's still a �, in 17 U.S. Code � 107 but I'm not sure what that symbol is supposed to be.

Also removed some stuff from the GitHub website which I accidently put in SW.README (and thought I had already removed).