Add Doonium and Durasteel 1x1 Round Wedge and remove normal variant
Also remove 1x1 Round Big strings which I forgot to remove previously.

Idk why there were 3 folders and sets of files for 1x1 Round Wedge armour, with seemingly 2 of them not doing anything (and not in the normal armour folder), but I removed all of them and that doesn't seem to have caused any problems.