Doonium price x 2/5 to be 2x normal armour price.
I know doonium is very expensive in actual Star Wars, but for this mod they're just not worth it for that high of a price. 2x normal armour for half the weight seems a lot more reasonably than 5x price (or 4.5 for some parts...).

Also fixed a durasteel price that was wrong.